+++ src/main/kotlin/io/kanro/idea/plugin/protobuf/aip/annotator/AipAnnotator.kt
@@ -5,6 +5,10 @@ import com.intellij.lang.annotation.AnnotationHolder
 import com.intellij.lang.annotation.Annotator
 import com.intellij.lang.annotation.HighlightSeverity
 import com.intellij.psi.PsiElement
+import io.kanro.idea.plugin.protobuf.aip.reference.AipResourceReference
+import io.kanro.idea.plugin.protobuf.aip.reference.ProtobufFieldReferenceInString
+import io.kanro.idea.plugin.protobuf.aip.reference.ProtobufRpcInputFieldReference
+import io.kanro.idea.plugin.protobuf.aip.reference.ProtobufRpcOutputFieldReference
 import io.kanro.idea.plugin.protobuf.lang.psi.ProtobufStringValue
 import io.kanro.idea.plugin.protobuf.lang.psi.ProtobufVisitor
 
@@ -14,14 +18,29 @@ class AipAnnotator : Annotator {
             override fun visitStringValue(o: ProtobufStringValue) {
                 o.reference?.let {
                     if (it.resolve() == null) {
-
-                        holder.newAnnotation(
-                            HighlightSeverity.ERROR,
-                            "Resource name ${o.text} not found."
-                        )
-                            .range(o.textRange)
-                            .highlightType(ProblemHighlightType.LIKE_UNKNOWN_SYMBOL)
-                            .create()
+                        when (it) {
+                            is AipResourceReference -> {
+                                holder.newAnnotation(
+                                    HighlightSeverity.ERROR,
+                                    "Resource name ${o.text} not found."
+                                )
+                                    .range(o.textRange)
+                                    .highlightType(ProblemHighlightType.LIKE_UNKNOWN_SYMBOL)
+                                    .create()
+                            }
+                            is ProtobufRpcInputFieldReference,
+                            is ProtobufRpcOutputFieldReference -> {
+                                holder.newAnnotation(
+                                    HighlightSeverity.ERROR,
+                                    "Field ${o.text} of message \"${
+                                    (it as ProtobufFieldReferenceInString).message()?.qualifiedName()
+                                    }\" not found."
+                                )
+                                    .range(o.textRange)
+                                    .highlightType(ProblemHighlightType.LIKE_UNKNOWN_SYMBOL)
+                                    .create()
+                            }
+                        }
                     }
                     return
                 }