+++ src/main/kotlin/io/kanro/idea/plugin/protobuf/lang/psi/primitive/ProtobufElementBase.kt
@@ -4,6 +4,7 @@ import com.intellij.extapi.psi.ASTWrapperPsiElement
 import com.intellij.lang.ASTNode
 import com.intellij.navigation.ItemPresentation
 import com.intellij.psi.PsiNameIdentifierOwner
+import io.kanro.idea.plugin.protobuf.lang.psi.primitive.feature.ProtobufNamedElement
 
 abstract class ProtobufElementBase(node: ASTNode) : ASTWrapperPsiElement(node) {
     override fun getPresentation(): ItemPresentation? {
@@ -19,5 +20,10 @@ abstract class ProtobufElementBase(node: ASTNode) : ASTWrapperPsiElement(node) {
             return this.nameIdentifier?.textOffset ?: super.getTextOffset()
         }
         return super.getTextOffset()
+    }
+
+    override fun getName(): String? {
+        if (this is ProtobufNamedElement) return name()
+        return null
     }
 }