+++ src/main/kotlin/io/kanro/idea/plugin/protobuf/lang/psi/primitive/feature/ProtobufNamedElement.kt
@@ -1,0 +1,7 @@
+package io.kanro.idea.plugin.protobuf.lang.psi.primitive.feature
+
+import io.kanro.idea.plugin.protobuf.lang.psi.primitive.ProtobufElement
+
+interface ProtobufNamedElement : ProtobufElement {
+    fun name(): String?
+}