+++ src/main/kotlin/io/kanro/idea/plugin/protobuf/lang/psi/primitive/structure/ProtobufValueAssign.kt
@@ -1,7 +1,6 @@
 package io.kanro.idea.plugin.protobuf.lang.psi.primitive.structure
 
 import io.kanro.idea.plugin.protobuf.lang.psi.ProtobufConstant
-import io.kanro.idea.plugin.protobuf.lang.psi.ProtobufFieldDefinition
 import io.kanro.idea.plugin.protobuf.lang.psi.findChild
 import io.kanro.idea.plugin.protobuf.lang.psi.primitive.ProtobufElement
 
@@ -11,5 +10,5 @@ interface ProtobufValueAssign : ProtobufElement {
         return findChild()
     }
 
-    fun field(): ProtobufFieldDefinition?
+    fun field(): ProtobufFieldLike?
 }