+++ src/main/kotlin/io/kanro/idea/plugin/protobuf/lang/reference/ProtobufEnumValueReference.kt
@@ -3,6 +3,7 @@ package io.kanro.idea.plugin.protobuf.lang.reference
 import com.intellij.openapi.util.TextRange
 import com.intellij.psi.PsiElement
 import com.intellij.psi.PsiReferenceBase
+import com.intellij.psi.impl.source.tree.LeafElement
 import io.kanro.idea.plugin.protobuf.lang.psi.ProtobufEnumValue
 import io.kanro.idea.plugin.protobuf.lang.psi.ProtobufEnumValueDefinition
 import io.kanro.idea.plugin.protobuf.lang.psi.enum
@@ -30,5 +31,10 @@ class ProtobufEnumValueReference(field: ProtobufEnumValue) :
             if (it !is ProtobufEnumValueDefinition) return@mapNotNull null
             (it as? ProtobufLookupItem)?.lookup()
         }?.toTypedArray() ?: arrayOf()
+    }
+
+    override fun handleElementRename(newElementName: String): PsiElement {
+        (element.identifierLiteral?.node as? LeafElement)?.replaceWithText(newElementName)
+        return element
     }
 }